CSV saving overhaul
Now it is a start/stop process. It starts writing to a timestamped file whatever there is in the buffer, and it is updated every time new data is added to the buffer. Gas name, decimal point and measurement units are appended as a header.

When the file is over 8MB a new one is created with a number suffi, because excel does not support files over 10MB so there is a safety margin.